move api calls to their own functions
