+++ src/api/getProfileInfo.js
@@ -1,0 +1,11 @@
+import axios from "axios";
+
+export const getProfileInfo = async (user) => {
+  return axios
+    .get(user.url, {
+      headers: {
+        Authorization: `token ${process.env.REACT_APP_GITHUB_ACCESS_TOKEN}`,
+      },
+    })
+    .then((res) => res.data);
+};
